+++ src/toolbox.ts
@@ -3,7 +3,7 @@ import type {
     FilterDistanceConfig,
     RecolorActiveConfig,
     ValidDeprecatedBy,
-} from "..";
+} from "../index";
 // Import ULabel from ../src/index - TypeScript will find ../src/index.d.ts for types
 import { ULabel } from "../src/index";
 import { DEFAULT_FILTER_DISTANCE_CONFIG } from "./configuration";